menambahkan fitur authentication & authorization

+++ Controllers/RoleController.cs
@@ -16,17 +16,29 @@ public class RoleController : Controller
 
     public IActionResult Index()
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var roles = roleRepository.GetAll();
         return View(roles);
     }
     public IActionResult Details(int id)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var role = roleRepository.GetById(id);
         return View(role);
     }
 
     public IActionResult Create()
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         return View();
     }
 
@@ -34,6 +46,10 @@ public class RoleController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Create(Role role)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var result = roleRepository.Insert(role);
         if (result > 0)
             return RedirectToAction(nameof(Index));
@@ -42,6 +58,10 @@ public class RoleController : Controller
 
     public IActionResult Edit(int id)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var role = roleRepository.GetById(id);
         return View(role);
     }
@@ -50,6 +70,10 @@ public class RoleController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Edit(Role role)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var result = roleRepository.Update(role);
         if (result > 0)
         {
@@ -60,6 +84,10 @@ public class RoleController : Controller
 
     public IActionResult Delete(int id)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var role = roleRepository.GetById(id);
         return View(role);
     }
@@ -68,6 +96,10 @@ public class RoleController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Remove(int id)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var result = roleRepository.Delete(id);
         if (result > 0)
         {
