+++ Controllers/UniversityController.cs
@@ -20,21 +20,25 @@ public class UniversityController : Controller
         {
             return RedirectToAction("Unauthorized", "Error");
         }
-        if (HttpContext.Session.GetString("role") != "Admin")
-        {
-            return RedirectToAction("Forbidden", "Error");
-        }
         var universities = repository.GetAll();
         return View(universities);
     }
     public IActionResult Details(int id)
     {
+        if (HttpContext.Session.GetString("email") == null)
+        {
+            return RedirectToAction("Unauthorized", "Error");
+        }
         var university = repository.GetById(id);
         return View(university);
     }
 
     public IActionResult Create()
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         return View();
     }
 
@@ -42,6 +46,10 @@ public class UniversityController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Create(University university)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var result = repository.Insert(university);
         if (result > 0)
             return RedirectToAction(nameof(Index));
@@ -50,6 +58,10 @@ public class UniversityController : Controller
 
     public IActionResult Edit(int id)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var university = repository.GetById(id);
         return View(university);
     }
@@ -58,6 +70,10 @@ public class UniversityController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Edit(University university)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var result = repository.Update(university);
         if (result > 0)
         {
@@ -68,6 +84,10 @@ public class UniversityController : Controller
 
     public IActionResult Delete(int id)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var university = repository.GetById(id);
         return View(university);
     }
@@ -76,6 +96,10 @@ public class UniversityController : Controller
     [ValidateAntiForgeryToken]
     public IActionResult Remove(int id)
     {
+        if (HttpContext.Session.GetString("role") != "Admin")
+        {
+            return RedirectToAction("Forbidden", "Error");
+        }
         var result = repository.Delete(id);
         if (result == 0)
         {
